wip(deferred): checkpoint the three deferred-work agent edits
Preserved before review, not because reviewed.
